Major changes. Handle some exceptions that should cause upload to fail. Added test coverage.

+++ ipm-be/csv-file-transfer-rest/src/main/java/com/kerneldc/ipm/rest/csv/service/ProcessingStats.java
@@ -7,11 +7,7 @@ import lombok.Data;
 @Builder
 public class ProcessingStats {
 
-	private Integer numberOfLinesInFile;
+	private Long numberOfLinesInFile;
 	private Integer numberOfExceptions;
 	private String elapsedTime;
-//	
-//	public void incrementExceptionsCounts(int count) {
-//		numberOfExceptions += count;
-//	}
 }
